week 3 third push

+++ src/week3/CastingPractice.java
@@ -5,7 +5,7 @@ public class CastingPractice {
     public static void main(String[] args) {
 
         int a = 10;
-        double b = a;
+        double b = (double) a;
 
         double c = 100;
         int d = (int)c;
